Replace make clean in zsh build script.
make clean got rid of the effects of changing the testdir but it is too slow.
Since the testdir is only used by modules and athame.c we can avoid the clean
by deleting zshpaths.h to rebuild the modules with the correct testdir and touching Src/Zle/athame.c so that it rebuilds with the correct testdir.