Change vim setup.
Vim setup is now spread out over init and loop. Athame will try to do as much as possible in init, but defers the rest to loop as soon as it reads a character.

This also fixes a bug introduced in the last commit breaking first_char code and a bug in detecting failed remote_exprs.