Get special keys from readline.
Get special keys from readline programatically instead of using the defaults.
This means that in python2 shell, which turns off tab completion, tabs go
directly to vim.

Note that python shells block on key inputs in a way that conflicts
with athame's tab completion hack. That is a different bug. It no longer
affects python2 because we bypass tab completion, but python3 shell actually
supports tab completion and so is still affected.