Catch --servername errors better
Previously, we wouldn't read the whole error string sometimes and give the user
the slower and more vague "Vimbed failure" error. Now, we check to see if vim
starts with "VIM", which is a hint something is wrong, and if so, sleep for 20
ms to accumulate more of the error (it's just an error, so it's pretty quick).