Added sanity check to tests.
On old versions of bash, we can't use negative arrray indexes. This is a
seperate problem, but more importantly, this error ends up causing the test
script to pass and allows the installation of Athame without running tests.
With the new sanity check, exiting runtest early for any reason counts as a test
failure.