Fix command mode display bug
Command mode wasn't getting displayed on update. This wasn't obvious except in
python because it was still changing from signals, which often happen in
command mode.

This was fixed by registering changes in command mode as changes. This was
previously avoided because of out-of-sync meta/text but that is no longer a
problem thanks to slices.