Fixed C-C for python.
Python is unable to handle SIGINTs that happen during readline code.
This is usually not a big deal because not much time is spent in
readline code, but Athame spends most of its time there. This is fixed
by a nastyish hack that catches the SIGINT and sends it back async to
the process.